bug fix.
In StAX, there's always a "next token" to the end element token. So when XSB is parsing a fragment, some implementations, in particular JAXB, tries to move beyond the end element. Today this fails with a XMLStreamReader created from a mark.

It's arguable whether moving beyond endElement was a good idea, but that's in the spec, and it's certainly allowed. Therefore, to improve the compatibility, I made a change so that it produces a virtual endDocument event. For a similar reason, it might be desirable to produce a dummy startDocument token.

Since we are particularly performance concious in this project, I chose to use an int (instead of enum) to represent the additional state. The additional performance penalty should be minimal, given that we were already checking the boolean flag.

svn path=/trunk/; revision=57